RavenDB-14706 Remove comments, move SetProviderSpecificDbType to SqlDatabaseWriter.cs, Add important annotaion for current development
